heroku/requirements.txt
requests==2.32.3
jinja2==3.1.4
matplotlib==3.9.2
weasyprint==62.3
apscheduler==3.10.4
python-dotenv==1.0.1
